Simplify the content off scrollphat.py and keep a full version.
